requirements.txt
django
pytest-playwright
#выполнить playwright install chromium
2captcha-python

